Allow setting custom image upload link
Before uploading to S3, we need to call an API function to generate a
signed PUT url. The link we need depends on the bucket we're uploading
to, so this is allowing different to configure which endpoint you call
to generate the link. This brings it in line with the previous file
upload comp.