Set pino log level at instance initialization, not just in the transports
Pino allows log levels to be set in the transports, but appears to
require that levels are set in the instance initialization as well as
the transports. None of the documentation that I saw made this clear,
and all documentation that I saw relating to Pino.Transport appears to
only set the level in the transport initialization. However, I have
tested this locally and in production and the fix works.