Stop country name from blocking event creation
Event country is a required field for new event creation, however if the
event is an online event, then the field for entering the country is
obscured. And the form does not subject (validation) but the error
message (on the field) is hidden.

This simple fix makes the instance country the default country for new
events, meaning that the country field is populated and the form does
not fail.