Feature/whatsapp send message (#82)
* Add gender to people schema

* Add required inputs to import script

* Run zapatos to reflect the latest migration

* Migration to people.people table to add gender and convert dob to a date

The view includes the dob field, so it needed to be deleted and
recreated in order to alter the existing column type.

* Updated auto translations

* Change date to be date type

* Create new date validation (using IsoDate (yyyy-mm-dd)

* New function to read tag by name, rather than ID

* Import tags and events from import spreadsheet

* Upated HTML-date compnent to accept datestrings rather than date types

* Update dob type and add gender dropdown

* Make sure the import object fits the createPerson schema

1. Strip out tags and events (these don't exist in the schema to create
a new person record)
2. Convert date_of_birth (in the spreadsheet) to dob

* Bump version of AWS s3 client to avoid a known bug

See https://github.com/aws/aws-sdk-js-v3/issues/6855 for more info

* Migration to enforce phone number and email uniqueness per instance

* Updated zapatos schema for new index

* [unrelated/chore] Get rid of MaybePromise from handler functions

MaybePromise was an old Sveltekit type equivalent to `Response |
Promise<Response>`. It's deprecated (for good reason) and all these
hooks can now take `Promise<Response>`

* Add whatsapp default phone number to env vars

* Clarify format of phone number

* Change all references to phone_number_id to become phone_number

This is what it was actually referring to, we shouldn't have been
calling it a phone_number_id when it's actually just a phone number.

The reason it was previously phone_number_id, is because that's the
identifier used by Meta's WhatsApp Business Cloud API. YCloud uses phone
numbers directly, so we should update the naming to reflect that.

* Change the instance CRUD functions to use phone_number instead of phone_number_id

* Set the initial default phone number for an instance

* Remove double import of paraglide messages

Not sure how this snuck in there

* Updated english localization strings

* Include the WamidID in updating sent message schema

This is important for ensuring we track the wamid (ie: the unique
message ID in meta's systems)

* Replace meta's biz_opaque_callback_data with YCloud's externalId

This is an abritrary string we can set to reference messages in our
system. externalId is the name used by yCloud

* Removed unused element from triggerAction schema

* Add clarifying comment to function that lacks tennant safety enforcement

* Convert event and petition IDs to numbers instead of strings

* Delete old WhatsApp webhook handler

This is for the WA business cloud API. We're using Ycloud so we don't
need it. IN the future, if we migrate back to Business Cloud we will
want to rewrite our current systems, not use this old code anyway

* Refactor the incoming whatsapp message hook handler

This is effectively a rewrite from scratch of this file in order to
achieve the aims of
https://www.notion.so/belcoda/RFC-001-2-Incoming-WhatsApp-messages-1c423474e3b380179735ccdcc0981a9b?pvs=4

As well as implementing our new strategy for deriving instances and
person info from incoming messsages, addition goals of the rewrite were
to make it simpler and much more readable, observable and testable, as
well as to make building in support for different types of messages
easier down the track.

It's a big change in a very important file, so I've tried to add lots of
comments and indications of the direction behind the decisions being
made.

* Remove unnecessary import

* Add some comments for future refactor

These are cases where parsing might seem like overkill because it's
being passed from one internal file to another with no external systems
in between. However, not parsing here removes our type safety, so
changes in the underlying schema won't be picked up by Typescript as
errors here. In the future, we should refactor this to be a union type
in order to cover all cases and still get the type safety we need.

* Fixed type/unclear comment

* Breaking out the received message handler into a separate queue message

This was previously implemented in the incoming message handler,
however, now breaking it out into multiple different handlers makes
these shared functions better in their own file

* Update semd message handler to properly set info for creating sent_whatsapp_message in database

* Remove generic incoming message queue handler

This is being split up into different handlers fo different message
types

* Create handlers for different message types that we currently support

* Remove old whatsapp incoming message handler

This is for the business cloud api, as we're using YCloud now we don't
need this code

* Rewrite of sent message update YCloud webhook handler

This was hastily copied over from the updated message event handler for
the Business Cloud API and then adapted for YCloud and it shows. This
rewrite handles the ycloud events in a much more readable and logical
flow. It also aims to be more testable.

* New upsert function for conversations

The WhatsApp conversation window can be updated with timing based on
when conversations begin and are updated. So an upsert function made
sense. However, with WhatsApp moving to a per-message pricing model,
it's unlikely we'll use this in reality. In the future, we might remove
the concept of "conversations" from the app entirely.

* New helper functions to get instance information from sent messages and wamid

* Autotranslated localization strings

* Change phone number to optional rather than nullable

This allows it to be not present in existing instance settings objects,
but still be created properly becuase it has a solid default value

* Change variable name to PHONE_NUMBER to clarify that it's a phone number not an ID

* Make DEFAULT_WHATSAPP_PHONE_NUMBER a public env var so it can be used in client side code (eg: validation schemas)

* Tidy up imports

* Improve logging and ensure that the response is always successful

If the response is not always 200, then the queue handler will retry the
send message job. Perhaps one day if we were sure that our code was
bulletproof that might be desired behaviour, but for now it is very
likely to cause more harm than good. And combined with good logging and
Sentry, we should be able to keep track of errors happening this way.

* Delete after_sent files

* Export a type required in send_message handler

* Refactor send_message handler and bring in post-sent processing logic into the same file

* Rename functions to not be confusing

These functions were called _getInstanceId... but actually returned a
full instance object. This was confusing and has now been fixed to read
getInstance

* Tests for incoming message handler

* Added missing ZH strings

* Added mocks to env vars to ensure tests run in CI

* Mocking another env var that is breaking CI tests

* Update the mocking methodology to use the method suggested in the error messages

* Fix pino import to not import other server modules which require env vars which aren't present in CI

* Tidied up imports to be inside test files (and after mocks are in effect)

* Cleanup formatting

* Split out send_message function to a separate action file

* Tests for send message actions

* Fix typo in migration sql