Remove alert from layout
This was an alert to make it clear when there wasn't a default send
signature. I then changed it to just fallback to
instanceSlug@belcoda.com if there is no send signature configured, and i
think that's fine. I'll leave this commented incase we need to come back
to it though