Update svelte-french-toast to use svelte5
We were previously using a fork of this toasting library, because it
didn't suppot svelte 5. But now they have released an alpha that has
svelte 5 support (finally) so we can switch back.

This has the side effect of fixing issues with broken icons, so the
toast messages should look good again.