Fix type errors in people file
Changing dob field to a date column has lead to type errors where
Zapatos expects either a Date or a string that is of the type
`${number}-${number}-${number}` instead of a string.

Even though the type it expects is a string, and valibot is parsing an
iso date (which is always YYYY-MM-DD) it was throwing erorrs.

Some type casting has been enough to fix it, but I've had to result to
//@ts-ignore in one place where no wrangling was getting there, because
fundamentally it works, the validation means its safe, and it's not
worth spending more time on.

However, I don't like having a //@ts-ignore at the point where Zapatos'
type guarantees would show us potential type errors. If that stays there
forever it will no doubt end up obscuring other bugs. We should be
careful, and remember about it if we're ever facing strange behaviour
with people uploads.