Remove the last remaining references to on: events in .svelte files (#112)
The last remaining component with on: event directives that I could find
(based on my searching) was the main button component (used very
widely).

This PR changes it to use an optional onclick() prop.

Most places that this was being imported were working fine anyway,
passing onclick which was being attached to either the <a> or <button>
element as a {...restProps}.

Only one place was actually passing on:click and that has been updated.

The only remaining on: directive is a carousel component which has no
file referneces and is not imported anywhere. I decided to keep it
incase it would break anything.