Feature/email from signatures (#110)
* Migration for from signaturs

* Some english localization messages

* An alert component to show a missing email from signature

This is to alert existing instances that they need to configure their
email signatures, because the system for sending has changed

* Update email related schemas for send signatures

* Update instances.settings schema for send signatures

* Remove localization t

* CRUD functions for email from signatures

* Remove localization t

* API routes for email from signatures

* Remove localization t

* Remove localization t

* CRON job handler for testing outstanding signature verification status

When a signature is submitted to Postmark, there isn't a webhook to tell
us when it has been verified. As such, we need to poll for status. This
is a cron job to do it on a 10 minute interval.

* Convert our send_event_notification cron job into a generic CRON job

It's triggered every 10 minutes and now includes the signature
verification trigger as well as the send event notification job

* Update zapatos schema

* CRUD routes in settings for email from signatures

* Add condition handling to ensure we have a workable send signature even if there isn't one set up

* Display alert if instance has no default from signature configured

* Update message schema (keeping old key)

* Delete route handler

* Add verified only listing

* Remove dedicated single-resource view route for from signatures

We can do everything we need to in the list view

* Build out list view

* Add delete CRUD db function

* Simplify signature select

Previously I had it loading via client side fetch, but now just loading
and inputing the server-side loaded send signatures as props

* Add links on settings page

* Set default currently selected send signature

Use the one that is already set in the email if exists, fallback to
default, fall back to null.

* Fix the queue processing jobs to actually select the correct dates

* Debug logging

* Remove unnecessary import

* Remove alert from layout

This was an alert to make it clear when there wasn't a default send
signature. I then changed it to just fallback to
instanceSlug@belcoda.com if there is no send signature configured, and i
think that's fine. I'll leave this commented incase we need to come back
to it though

* Localization strings

* Fix up issue with Error being imported as a component

This clashes with the standard `new Error()` in js

* Fix type in link

* Added alert about public email addresses