Added petition delete
I found that we didn't have a del function for petitions? Or maybe it
got lost in a merge conflict resolution issue or something? Either way,
the compiler was showing errors when we try to import del from this file
in the +server.ts route for deleting petitions. Everything else seems to
be working fine in terms of the petition soft delete functions.