Update domain validation schema
This schema previously accepted domain namae looking strings (eg:
example.com) or fully formed URLs (eg: https://example.com) and output
them as-is.

This caused issues with the onboarding forms because the instance
creation script expects the home page URL for an instance to be a fully
formed URL.

This has now been updated to attempt to transform incoming strings into
fully form URLs and output the URLs. It should only throw if the string
cannot be coerced into a URL (ie: using new URL())

This commit also includes a number of test cases for this validation
function, attempting to cover the kinds of URLs/domains that people
might input into a form.