Silence compiler warnings with dummy init functions
Sveltekit compiler wants to see init functions across client and server
in some instances.

In this case, I believe it's because we have a Sentry init function, but
no explicit init hook.

This fix was recommended by one of the on a github issue
(https://github.com/sveltejs/kit/discussions/12737) and seems to solve
the problem, although it feels like a bit of a dumb hack.

I can't see it impacting anything and it's an empty function that fires
once on server startup so it won't have any performance hit.