Patch/circular deps (#122)
* Add vite build-time circular dependency detector plugin

* Fix circular dependencies in $lib/server

* FIx circular dependencies in $lib/server/api/admins

This one required breaking out some logic into smaller modules imported
by both files, rather than having them import from each other.

* Add a comment noting that the circular dep is a type only import

This should not impact the vite bundler at all, so should be okay to
retain

* Resolved circular depenencies in groups/interactions

This involved splitting out a couple of functions into their own
files/modules in order to remove the need for any circular dependencies
between files

* Added @circular-ignore tag to avoid build errors

* Remove unused exports

These generate compiler warnings, but probably aren't related to the dev
server slowdowns. Still, I thought it was worth doing

* Change context="module" to module in module exports

This is due to a change in the svelte module context. Context="module"
is deprecated and throws compiler warnings. Probably not a big deal, but
can't hurt.

* Added petition delete

I found that we didn't have a del function for petitions? Or maybe it
got lost in a merge conflict resolution issue or something? Either way,
the compiler was showing errors when we try to import del from this file
in the +server.ts route for deleting petitions. Everything else seems to
be working fine in terms of the petition soft delete functions.

* Removed unnecessary imports from valibot module

* Remove unnecessary imports from '@internationalized/date'

* Comment to ignore circular import of svelte component importing itself

* Comment to ignore circular imports in shadcn-svelte components

This probably isn't ideal, and if we keep having issues we can come back
and attempt to fix it

* Update the //@circular-ignore comments to the correct syntax ;)

* Silence compiler warnings with dummy init functions

Sveltekit compiler wants to see init functions across client and server
in some instances.

In this case, I believe it's because we have a Sentry init function, but
no explicit init hook.

This fix was recommended by one of the on a github issue
(https://github.com/sveltejs/kit/discussions/12737) and seems to solve
the problem, although it feels like a bit of a dumb hack.

I can't see it impacting anything and it's an empty function that fires
once on server startup so it won't have any performance hit.

* Remove @valibot/to-json-schema module

It has a number of circular dependencies and wasn't being used for
anything (the plan was to use it for autodocumenting Swagger API specs),
but I don't think we'll need that for a long time.

* Removed zod from dependencies

It is still in there because it's an optional dependency of
sveltekit-superforms and openai

* Add circular dependencies output to .gitignore

* Added explicit flag about uploading sentry source maps

We do want to autoupload source maps in most instances, but it can be
useful to disable this when attempting to resolve compiler issues,
because it generates a huge amount of built-time noise to stdout