Bugfix/people imports (#123)
* improve parsing of gender

This accepts also "m" and "f" as valid. Anything unrecognised is set as "not_specified"

* Improve parsing of preferred language

Field is trimmed first to remove empty spaces that fail the 2-character validation. Defaults to instance language

* default dob to null if it's not a valid ISO date

* account for countries that aren't 2-character ISO codes

We will default to instance country if we cannot parse it as a two-letter ISO code

* skip rows that are entirely empty or don't have names

* Disable strict phone number validation

* remove events and tags from person object to be inserted

* csv helper class

* test cases

* refractor to make logic more testable

* refractoring. move a bunch of stuff around to different methods and files

Mean to make testing easier. Moved functions to util files rather than the server handlers which are harder to test

* Removed organization names

* exclude failed_rows_details from being saved in the database

* CSVs exported from Google Sheets will represent boolean vals as "TRUE" or "FALSE" (uppercase)

* Fix country defaulting to instance country

This was happening because we were comparing upper case strings to lower case ones

* Allow custom non-null languages

---------

Co-authored-by: Django Merope <dmerope@gmail.com>